Class for storing IP addresses

ip_address() constructs a vector of IP addresses.

is_ip_address() checks if an object is of class ip_address.

ip_address(ip = character())

is_ip_address(x)

# S3 method for ip_address
format(x, ...)

# S3 method for ip_address
as.character(x, ...)

Arguments

ip
Character vector of IP addresses, in dot-decimal notation (IPv4) or hexadecimal notation (IPv6).

x
An ip_address vector

...
Arguments to be passed to other methods

Value

An ip_address vector

Details

An address in IPv4 space uses 32-bits. It is usually represented as 4 groups of 8 bits, each shown as decimal digits (e.g. "192.168.0.1"). This is known as dot-decimal notation.

An address in IPv6 space uses 128-bits. It is usually represented as 8 groups of 16 bits, each shown as hexadecimal digits (e.g. "2001:0db8:85a3:0000:0000:8a2e:0370:7334"). This representation can also be compressed by removing leading zeros and replacing consecutive groups of zeros with double-colon (e.g. "2001:db8:85a3::8a2e:370:7334").

The ip_address() constructor accepts a character vector of IP addresses in these two formats. It checks whether each string is a valid IPv4 or IPv6 address, and converts it to an ip_address object. If the input is invalid, a warning is emitted and NA is stored instead.

When casting an ip_address object back to a character vector using as.character(), IPv6 addresses are reduced to their compressed representation.

This class also supports bitwise operations: & (AND), | (OR) and ! (NOT).

Examples

# supports IPv4 and IPv6 simultaneously
ip_address(c("0.0.0.1", "192.168.0.1", "2001:db8::8a2e:370:7334"))

#> <ip_address[3]>
#> [1] 0.0.0.1                 192.168.0.1             2001:db8::8a2e:370:7334

# validates inputs and replaces with NA
ip_address(c("1.2.3.4", "255.255.255.256", "1.2.3.4/5"))

#> Warning: Invalid input in row 2: 255.255.255.256

#> Warning: Invalid input in row 3: 1.2.3.4/5

#> <ip_address[3]>
#> [1] 1.2.3.4 <NA>    <NA>   

# bitwise NOT
!ip_address("192.168.0.1")

#> <ip_address[1]>
#> [1] 63.87.255.254

# bitwise AND
ip_address("192.168.0.1") & ip_address("255.0.0.255")

#> <ip_address[1]>
#> [1] 192.0.0.1

# bitwise OR
ip_address("192.168.0.0") | ip_address("0.0.0.1")

#> <ip_address[1]>
#> [1] 192.168.0.1