Class for storing IP networks

ip_network() constructs a vector of IP networks.

is_ip_network() checks if an object is of class ip_network.

ip_network(ip = character(), strict = TRUE)

is_ip_network(x)

# S3 method for ip_network
format(x, ...)

# S3 method for ip_network
as.character(x, ...)

Arguments

ip
Character vector of IP networks, in CIDR notation (IPv4 or IPv6).

strict
If strict = TRUE (the default) and the input has host bits set, then a warning is emitted and NA is returned. If FALSE, the host bits are set to zero and a valid IP network is returned.

x
An object

...
Additional arguments to be passed to or from methods

Details

An IP network corresponds to a range of contiguous IP addresses (also known as an IP block). CIDR notation represents an IP network as the routing prefix address (which denotes the start of the range) and the prefix length (which indicates the size of the range) separated by a forward slash. For example, "192.168.100.0/22" represents addresses from "192.168.100.0" to "192.168.103.255".

The prefix length indicates the number of bits reserved by the routing prefix. This means that larger prefix lengths indicate smaller networks. The maximum prefix length is 32 for IPv4 and 128 for IPv6. These would correspond to an IP network of a single IP address.

The ip_network() constructor accepts a character vector of IP networks in CIDR notation. It checks whether each string is a valid IPv4 or IPv6 network, and converts it to an ip_network object. If the input is invalid, a warning is emitted and NA is stored instead.

When casting an ip_network object back to a character vector using as.character(), IPv6 addresses are reduced to their compressed representation.

See also

Examples

# supports IPv4 and IPv6 simultaneously
ip_network(c("92.0.2.0/24", "192.168.100.0/22", "2001:db8::/48"))

#> <ip_network[3]>
#> [1] 92.0.2.0/24      192.168.100.0/22 2001:db8::/48   

# validates inputs and replaces with NA
ip_network(c("192.168.0.0/24", "192.168.0.0/33", "1.2.3.4"))

#> Warning: Invalid argument: 192.168.0.0/33

#> Warning: Invalid argument: 1.2.3.4

#> <ip_network[3]>
#> [1] 192.168.0.0/24 <NA>           <NA>          

# IP networks should not have any host bits set
ip_network("192.168.0.1/22")

#> Warning: Invalid argument: 192.168.0.1/22 has host bits set

#> <ip_network[1]>
#> [1] <NA>

ip_network("192.168.0.1/22", strict = FALSE)

#> <ip_network[1]>
#> [1] 192.168.0.0/22