List constituent networks of an address range

Given an address range, this returns the list of constituent networks.

If you know the address range matches the boundaries of a single network, it might be preferable to use common_network(). This returns an ip_network vector instead of a list of ip_network vectors.

Usage

summarize_address_range(address1, address2)

Arguments

address1

An ip_address vector

address2

An ip_address vector

Value

A list of ip_network vectors

See also

Examples

# address range matches network boundaries
summarize_address_range(ip_address("192.168.0.0"), ip_address("192.168.0.15"))
#> <list_of<ip_network>[1]>
#> [[1]]
#> <ip_network[1]>
#> [1] 192.168.0.0/28
#> 

# address range does not match network boundaries
summarize_address_range(ip_address("192.167.255.255"), ip_address("192.168.0.16"))
#> <list_of<ip_network>[1]>
#> [[1]]
#> <ip_network[3]>
#> [1] 192.167.255.255/32 192.168.0.0/28     192.168.0.16/32   
#>