Vector of IP networks

ip_network() constructs a vector of IP networks.

as_ip_network() casts an object to ip_network.

is_ip_network() checks if an object is of class ip_network.

ip_network(...)

# S3 method for default
ip_network(ip = character(), strict = TRUE, ...)

# S3 method for ip_address
ip_network(address, prefix_length, strict = TRUE, ...)

as_ip_network(x)

is_ip_network(x)

# S3 method for ip_network
format(x, ...)

# S3 method for ip_network
as.character(x, ...)

Arguments

...
Arguments to be passed to other methods

ip
Character vector of IP networks, in CIDR notation (IPv4 or IPv6).

strict
If TRUE (the default) and the input has host bits set, then a warning is emitted and NA is returned. If FALSE, the host bits are set to zero and a valid IP network is returned. If you need to retain the host bits, consider using ip_interface() instead.

address
An ip_address vector

prefix_length
An integer vector

x
An ip_network vector

Value

An ip_network vector

Details

An IP network corresponds to a range of contiguous IP addresses (also known as an IP block). CIDR notation represents an IP network as the routing prefix address (which denotes the start of the range) and the prefix length (which indicates the size of the range) separated by a forward slash. For example, 192.168.100.0/22 represents addresses from 192.168.100.0 to 192.168.103.255.

The prefix length indicates the number of bits reserved by the routing prefix. This means that larger prefix lengths indicate smaller networks. The maximum prefix length is 32 for IPv4 and 128 for IPv6. These would correspond to an IP network of a single IP address.

The ip_network() constructor accepts a character vector of IP networks in CIDR notation. It checks whether each string is a valid IPv4 or IPv6 network, and converts it to an ip_network object. If the input is invalid, a warning is emitted and NA is stored instead.

An alternative constructor accepts an ip_address vector and an integer vector containing the network address and prefix length, respectively.

When casting an ip_network object back to a character vector using as.character(), IPv6 addresses are reduced to their compressed representation.

See also

Examples

# construct from character vector
ip_network(c("192.168.0.0/24", "2001:db8::/48"))

#> <ip_network[2]>
#> [1] 192.168.0.0/24 2001:db8::/48 

# construct from address + prefix length objects
ip_network(ip_address(c("192.168.0.0", "2001:db8::")), c(24L, 48L))

#> <ip_network[2]>
#> [1] 192.168.0.0/24 2001:db8::/48 

# validates inputs and replaces with NA
ip_network(c("192.168.0.0/33", "192.168.0.0"))

#> Warning: Invalid value on row 1: 192.168.0.0/33

#> Warning: Invalid value on row 2: 192.168.0.0

#> <ip_network[2]>
#> [1] <NA> <NA>

# IP networks should not have any host bits set
ip_network("192.168.0.1/22")

#> Warning: Invalid value on row 1: 192.168.0.1/22 (host bits set)

#> <ip_network[1]>
#> [1] <NA>

# but we can mask the host bits if desired
ip_network("192.168.0.1/22", strict = FALSE)

#> <ip_network[1]>
#> [1] 192.168.0.0/22